+++ packages/scripts/src/cli.ts
@@ -8,43 +8,39 @@ import { Command } from "commander";
 import { runBuild } from "./commands/build";
 import { ALL_PACKAGES, CATEGORY_VM } from "./common/cli.constants";
 import { startDeleteFlow } from "./commands/delete";
-import {
-  Options_Cli,
-  Options_Cli_Build,
-  Options_Cli_Delete,
-  Schema_Options_Cli_Build,
-  Schema_Options_Cli_Delete,
-  Schema_Options_Cli_Root,
-} from "./common/cli.types";
-import { getPckgsTo, log } from "./common/cli.utils";
+import { CliValidator } from "./cli.validator";
 
 class CompassCli {
   private program: Command;
-  private options: Options_Cli;
+  private validator: CliValidator;
 
   constructor(args: string[]) {
     this.program = this._createProgram();
+    this.validator = new CliValidator(this.program);
     this.program.parse(args);
-    this.options = this._getCliOptions();
   }
 
   public async run() {
-    const { force, user } = this.options;
+    const options = this.validator.getCliOptions();
+    const { force, user } = options;
     const cmd = this.program.args[0];
 
     switch (true) {
       case cmd === "build": {
-        await this._validateBuild();
-        await runBuild(this.options);
+        await this.validator.validateBuild(options);
+        await runBuild(options);
         break;
       }
       case cmd === "delete": {
-        this._validateDelete();
+        this.validator.validateDelete(options);
         await startDeleteFlow(user as string, force);
         break;
       }
       default:
-        this._exitHelpfully("root", `${cmd as string} is not a supported cmd`);
+        this.validator.exitHelpfully(
+          "root",
+          `${cmd as string} is not a supported cmd`
+        );
     }
   }
 
@@ -77,156 +73,6 @@ class CompassCli {
         "specify which user to run script for"
       );
     return program;
-  }
-
-  private _exitHelpfully(cmd: "root" | "build" | "delete", msg?: string) {
-    msg && log.error(msg);
-
-    if (cmd === "root") {
-      console.log(this.program.helpInformation());
-    } else {
-      const command = this.program.commands.find(
-        (c) => c.name() === cmd
-      ) as Command;
-      console.log(command.helpInformation());
-    }
-
-    process.exit(1);
-  }
-
-  private _getBuildOptions() {
-    const buildOpts: Options_Cli_Build = {};
-
-    const buildCmd = this.program.commands.find(
-      (cmd) => cmd.name() === "build"
-    );
-    if (buildCmd) {
-      const packages = this.program.args[1]?.split(",");
-      if (packages) {
-        buildOpts.packages = packages;
-      }
-
-      const environment = buildCmd?.opts()[
-        "environment"
-      ] as Options_Cli_Build["environment"];
-      if (environment) {
-        buildOpts.environment = environment;
-      }
-
-      const clientId = buildCmd?.opts()[
-        "clientId"
-      ] as Options_Cli_Build["clientId"];
-      if (clientId) {
-        buildOpts.clientId = clientId;
-      }
-    }
-    return buildOpts;
-  }
-
-  private _getCliOptions(): Options_Cli {
-    const options = this._mergeOptions();
-    const validOptions = this._validateOptions(options);
-
-    console.log("options", options);
-    console.log("validOptions:", validOptions);
-    return validOptions;
-  }
-
-  private _getDeleteOptions() {
-    const deleteOpts: Options_Cli_Delete = {};
-
-    const deleteCmd = this.program.commands.find(
-      (cmd) => cmd.name() === "delete"
-    );
-    if (deleteCmd) {
-      const user = deleteCmd?.opts()["user"] as Options_Cli["user"];
-      if (user) {
-        deleteOpts.user = user;
-      }
-    }
-
-    return deleteOpts;
-  }
-
-  private _mergeOptions = (): Options_Cli => {
-    const _options = this.program.opts();
-    let options: Options_Cli = {
-      ..._options,
-      force: _options["force"] === true,
-    };
-
-    const buildOptions = this._getBuildOptions();
-    if (Object.keys(buildOptions).length > 0) {
-      options = {
-        ...options,
-        ...buildOptions,
-      };
-    }
-
-    const deleteOptions = this._getDeleteOptions();
-    if (Object.keys(deleteOptions).length > 0) {
-      options = {
-        ...options,
-        ...deleteOptions,
-      };
-    }
-
-    return options;
-  };
-
-  private async _validateBuild() {
-    if (!this.options.packages) {
-      this.options.packages = await getPckgsTo("build");
-    }
-
-    const unsupportedPackages = this.options.packages.filter(
-      (pkg) => !ALL_PACKAGES.includes(pkg)
-    );
-    if (unsupportedPackages.length > 0) {
-      this._exitHelpfully(
-        "build",
-        `One or more of these packages isn't supported: ${unsupportedPackages.toString()}`
-      );
-    }
-  }
-
-  private _validateDelete() {
-    const { user } = this.options;
-    if (!user || typeof user !== "string") {
-      this._exitHelpfully("delete", "You must supply a user");
-    }
-  }
-
-  private _validateOptions(options: Options_Cli) {
-    const { data: rootData, error: rootError } =
-      Schema_Options_Cli_Root.safeParse(options);
-    if (rootError) {
-      this._exitHelpfully(
-        "root",
-        `Invalid CLI options: ${rootError.toString()}`
-      );
-    }
-
-    const { data: buildData, error: buildError } =
-      Schema_Options_Cli_Build.safeParse(options);
-    if (buildError) {
-      this._exitHelpfully(
-        "build",
-        `Invalid build options: ${buildError.toString()}`
-      );
-    }
-
-    const { data: deleteData, error: deleteError } =
-      Schema_Options_Cli_Delete.safeParse(options);
-    if (deleteError) {
-      this._exitHelpfully(
-        "delete",
-        `Invalid delete options: ${deleteError.toString()}`
-      );
-    }
-
-    const data: Options_Cli = { ...rootData, ...buildData, ...deleteData };
-    return data;
   }
 }
